This also drops some changes which should be reworked:
  * DefaultIdeaErrorHandler.java
  * resources/src/idea/JavaActions.xml

Bazel changes were dropped similar to last merge and can be looked into
later.

Additionally, dropping our ConfigImportHelperTest as we do not run these
and we will only maintain our patch here for N.1 and N.2 release.

Bug: 397701479

+++ notebooks/visualization/src/com/intellij/notebooks/visualization/ui/EditorEmbeddedComponentLayoutManager.kt
@@ -74,7 +74,7 @@ internal class EditorEmbeddedComponentLayoutManager(private val editor: EditorEx
       val newBounds = Rectangle(
         inlayBounds.x,
         inlayBounds.y,
-        min((if (constraint.isFullWidth) visibleWidth else size.width).toDouble(), visibleWidth.toDouble()).toInt(),
+        min((if (constraint.isFullWidth) visibleWidth else size.width), visibleWidth),
         size.height
       )
       if (newBounds != component.bounds) {
